requirements.txt
pytest==7.4.0
pytest-cov==4.0.0
black==22.3.0
pylint==2.15.3

pandas==2.2.2
matplotlib==3.9.2
sqlalchemy==2.0.35

click==8.1.3 
requests==2.31.0
# add necessary databricks req
python-dotenv
databricks-sql-connector
